+++ src/main/java/insane96mcp/iguanatweaksexpanded/integration/emi/EmiMultiItemSmeltingRecipe.java
@@ -40,7 +40,7 @@ public abstract class EmiMultiItemSmeltingRecipe implements EmiRecipe {
 		this.recycle = recipe.getRecycle();
 		int amount = (int) (this.outputIncrease + 1);
 		if (this.recycle != null)
-			amount = this.recycle.amountAtFullDurability();
+			amount = (int) (this.recycle.amountAtFullDurability() * this.recycle.ratio());
 		this.output = EmiStack.of(recipe.getResult(), (long) (amount));
 		this.cookingTime = recipe.getCookingTime();
 	}
@@ -82,7 +82,12 @@ public abstract class EmiMultiItemSmeltingRecipe implements EmiRecipe {
 		widgets.addText(Component.translatable("emi.cooking.experience", this.experience), 92, 5, -1, true);
 		SlotWidget outputSlot = widgets.addSlot(this.output, 125, 12).large(true).recipeContext(this);
 		if (this.recycle != null)
-			outputSlot.appendTooltip(Component.translatable("emi.tooltip.iguanatweaksexpanded.recycle", InsaneLib.ONE_DECIMAL_FORMATTER.format(this.recycle.ratio() * 100f)).withStyle(ChatFormatting.DARK_GREEN));
+			outputSlot.appendTooltip(Component.translatable("emi.tooltip.iguanatweaksexpanded.recycle").withStyle(ChatFormatting.DARK_GREEN));
+		if (this.outputIncrease > 0f) {
+			float chance = this.outputIncrease - (int) this.outputIncrease;
+			if (chance > 0f)
+				outputSlot.appendTooltip(Component.translatable("emi.tooltip.iguanatweaksexpanded.output_increase", InsaneLib.ONE_DECIMAL_FORMATTER.format(chance * 100f)).withStyle(ChatFormatting.DARK_GREEN));
+		}
 		/*widgets.addSlot(this.ingredient, 2, 2);
 		widgets.addSlot(this.gear, 2, 38);
 		widgets.addText(Component.literal("" + this.smashesRequired), 32, 11, 0xFFFFFF, true);