+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Fixed starting season config setting the previous subseason
+* Fixed wrong enchantability on some materials
 
 ## 1.7.0 - The "The recipe book sucks" update
 * Emi support!