+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## 1.7.1
 * Fixed starting season config setting the previous subseason
 * Fixed wrong enchantability on some materials